Add files via upload
I was inspired to create this website when my my uncle's family( with kids aged 5 and 10) moved to a city in India from a different country and found it difficult to adjust to the city culture. Having been non-residents of the city for 20 years, they found it hard to adjust to the lifestyle, time, management, move, etc. Moving to a new city can be exciting given you have plenty of time to adjust and have minimal responsibilities. Having to cope mentally with the burdens of moving all you belongings and establishing yourself with your utilities, banks, rental agreements, etc is hard. Every city has its own essence, it's own culture, most of which is subtly embedded. Recognizing this could ease the move greatly. This website aims to do exactly that. 

With the websites, services, the user can access some services such as groceries, time, money, friends, helping services, etc just by enrolling! 

Makes moving exciting again.